Restructured the 'connectors' by splitting them up. Easier to do this now than later, but trivial to glue them together in the future, so this is what I'll do until another way recommends itself.
