+++ imports/api/postgresql/async-utilities/comments.js
@@ -1,0 +1,14 @@
+import { pool } from "../connection-pool.js";
+
+export default class CommentConnector {
+	/*
+		getCommentById(id)
+
+		getAllComments(skip,limit)
+		getCommentsByAuthor(id,skip,limit)
+
+		writeComment
+		editComment
+		deleteComment
+	*/
+}
