+++ imports/api/postgresql/async-utilities/jobads.js
@@ -1,0 +1,13 @@
+import { pool } from "../connection-pool.js";
+
+export default class JobAdConnector {
+	/*
+		getJobAdById(id)
+		getAllJobAds(skip,limit)
+		getJobAdsByCompany(companyName,skip,limit)
+
+		postJobAd
+		editJobAd
+		deleteJobAd
+	*/
+}
