+++ imports/api/postgresql/async-utilities/reviews.js
@@ -1,0 +1,18 @@
+import { pool } from "../connection-pool.js";
+
+export default class ReviewConnector {
+	/*
+		getReviewById(id)
+
+		getReviewsByAuthor(id,skip,limit)
+		getAllReviews(skip,limit) -> gets all reviews, and assembles all
+							location and voting info
+		getReviewsForCompany(name,skip,limit) -> gets all reviews for company,
+							assembles all voting and location info
+
+		submitReview
+		editReview
+		deleteReview
+
+	*/
+}
