+++ imports/api/postgresql/async-utilities/salaries.js
@@ -1,0 +1,15 @@
+import { pool } from "../connection-pool.js";
+
+export default class SalaryConnector {
+	/*
+		getSalaryById
+
+		getAllSalaries(skip,limit)
+		getSalariesByAuthor(id,skip,limit)
+		getSalariesForCompany(name,skip,limit)
+
+		submitSalary
+		editSalary
+		deleteSalary
+	*/
+}
