+++ imports/api/postgresql/async-utilities/votes.js
@@ -1,0 +1,14 @@
+import { pool } from "../connection-pool.js";
+
+export default class VoteConnector {
+	/*
+		getVoteById(id) -> currently not possible, does it even make sense?
+
+		getAllVotes(skip,limit)
+		getVotesByAuthor(id,skip,limit)
+		getVotesForSubject(subject,refersto,skip,limit)
+
+		castVote
+		removeVote
+	*/
+}
